+++ src/app/d3components/utilities/trendLine.js
@@ -9,7 +9,9 @@ class TrendLine extends React.Component {
 
   constructor(props) {
     super(props);
-    this.componentWillMount = this.componentWillMount.bind(this);
+    this.updateSize = this.updateSize.bind(this);
+    this.minMaxing = this.minMaxing.bind(this);
+    this.getEndPoints = this.getEndPoints.bind(this);
     this.state = {
       width: this.props.width,
       data: []
@@ -17,12 +19,11 @@ class TrendLine extends React.Component {
   }
 
   componentWillMount() {
-    const _self = this;
     window.addEventListener('resize', function() {
-      _self.minMaxing();
-      _self.updateSize();
-      _self.getEndPoints();
-    }, true);
+      this.minMaxing;
+      this.updateSize;
+      this.getEndPoints;
+    }, false);
     this.setState({width: this.props.width});
   }
 
@@ -31,16 +32,24 @@ class TrendLine extends React.Component {
   }
 
   componentWillUnmount() {
-    const _self = this;
     window.removeEventListener('resize', function() {
-      _self.minMaxing();
-      _self.updateSize();
-    });
+      this.minMaxing;
+      this.updateSize;
+      this.getEndPoints;
+    }, false);
+  }
+
+  updateSize() {
+    let node = ReactDOM.findDOMNode(this);
+    let parentWidth = node.offsetWidth;
+    (parentWidth < this.props.width) ? 
+      this.setState({width:parentWidth}) :
+      this.setState({width:this.props.width});
   }
 
   repaintComponent() {
     const _self = this;
-    const forceResize = function(){
+    const forceResize = function() {
       _self.minMaxing();
       _self.updateSize();
       _self.getEndPoints();
@@ -53,29 +62,19 @@ class TrendLine extends React.Component {
     onRepaint(forceResize);
   }
 
-  updateSize(){
-    let node = ReactDOM.findDOMNode(this);
-    let parentWidth = node.offsetWidth;
-    if (parentWidth < this.props.width) {
-      this.setState({width:parentWidth});
-    } else {
-      this.setState({width:this.props.width});
-    }
-  }
-
   createChart(_self) {
 
     // Create line
     this.line = d3.svg.line()
       .x(function (d) {
-        return this.props.x(d[_self.props.xData]);
+        return this.props.x(d[_self.props.xDataKey]);
       })
       .y(function (d) {
-        return this.props.y(d[_self.props.yData]);
+        return this.props.y(d[_self.props.yDataKey]);
       });
 
     this.dataNest = d3.nest()
-        .key(function(d) {return d.type;})
+        .key(function(d) {return d.label;})
         .entries(this.state.data);
   }
 
@@ -110,8 +109,8 @@ class TrendLine extends React.Component {
 
     let minMax = {};
     let lineData = this.props.lineExtend;
-    minMax['xMin'] = d3.min(lineData.map(function (d) { return d[_self.props.xData]; }));
-    minMax['xMax'] = d3.max(lineData.map(function (d) { return d[_self.props.xData]; }));
+    minMax['xMin'] = d3.min(lineData.map(function (d) { return d[_self.props.xDataKey]; }));
+    minMax['xMax'] = d3.max(lineData.map(function (d) { return d[_self.props.xDataKey]; }));
 
     return minMax;
   }
@@ -127,8 +126,8 @@ class TrendLine extends React.Component {
       data = this.props.data;
     }
 
-    let xSeries = data.map(function (d) { return d[_self.props.xData]; });
-    let ySeries = data.map(function (d) { return d[_self.props.yData]; });
+    let xSeries = data.map(function (d) { return d[_self.props.xDataKey]; });
+    let ySeries = data.map(function (d) { return d[_self.props.yDataKey]; });
 
     let leastSquaresCoeff = this.leastSquares(xSeries, ySeries);
     let trendExtend = this.minMaxing();
@@ -140,12 +139,12 @@ class TrendLine extends React.Component {
 
     let trendData = [
       {
-        "type": "Trend Line",
+        "label": "Trend Line",
         'x':x1,
         'y':y1
       },
       {
-        "type": "Trend Line",
+        "label": "Trend Line",
         'x':x2,
         'y':y2
       }
@@ -188,10 +187,8 @@ TrendLine.propTypes = {
   lineNumbers: React.PropTypes.oneOf(['single','multi']),
   x: React.PropTypes.func,
   y: React.PropTypes.func,
-  xData: React.PropTypes.string.isRequired,
-  yData: React.PropTypes.string.isRequired,
-  xAxisLabel: React.PropTypes.string,
-  yAxisLabel: React.PropTypes.string,
+  xDataKey: React.PropTypes.string.isRequired,
+  yDataKey: React.PropTypes.string.isRequired,
   margin: React.PropTypes.object,
   yMaxBuffer: React.PropTypes.number
 };